feature_engineering/cloud_functions/requirements.txt
google-cloud-storage
google-cloud-datastore   
scikit-learn==0.20.3
opencv-python==4.0.0.21
click==7.0
numpy==1.16.2
pandas==0.24.2
xgboost==0.82
